requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.0
h11==0.14.0
idna==3.10
joblib==1.4.2
matplotlib==3.8.0
numpy==1.26.4
pandas==2.2.2
pydantic==2.11.3
pydantic_core==2.33.1
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.30.6
